# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Calendar Api",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:5000/api"
        }
    ],
    "paths": {
...
        "/admin/calendar?date=": {
            "get": {
                "tags": [
                    "Calendar"
                ],
                "summary": "Get calendar data",
                "description": "Get all events for specified month",
                "operationId": "calendar",
                "parameters": [
                    {
                        "name": "date",
                        "in": "query",
                        "description": "Calendar data",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "date=2024-09-30": {
                                "summary": "Data of all months' events",
                                "value": "2024-09-30"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "Data array",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "month_data": {
                                                        "type": "array",
                                                        "items": {
                                                            "properties": {
                                                                "id": {
                                                                    "description": "ID of the Event",
                                                                    "type": "integer",
                                                                    "example": "1"
                                                                },
                                                                "title": {
                                                                    "description": "Title of the Event",
                                                                    "type": "string",
                                                                    "example": "Title of the Event"
                                                                },
                                                                "start": {
                                                                    "description": "Time when Event have to start",
                                                                    "type": "string",
                                                                    "example": "10:30"
                                                                },
                                                                "duration": {
                                                                    "description": "Duration of the Event (hrs)",
                                                                    "type": "string",
                                                                    "example": "1:30"
                                                                },
                                                                "type_id": {
                                                                    "description": "ID of the Event's type",
                                                                    "type": "integer",
                                                                    "example": "1"
                                                                },
                                                                "status_id": {
                                                                    "description": "ID of the Event's status",
                                                                    "type": "integer",
                                                                    "example": "1"
                                                                },
                                                                "description": {
                                                                    "description": "Short Description of the Event",
                                                                    "type": "string",
                                                                    "example": "Event description"
                                                                },
                                                                "date": {
                                                                    "description": "The date to which the event is linked",
                                                                    "type": "string",
                                                                    "example": "2024-09-04 18:16:00"
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    },
                                                    "base_date": {
                                                        "description": "Date of the month",
                                                        "type": "string",
                                                        "example": "2024-08-31 02:10:11"
                                                    },
                                                    "month": {
                                                        "description": "Month for displaying events",
                                                        "type": "string",
                                                        "example": "August"
                                                    },
                                                    "year": {
                                                        "description": "Year for displaying events",
                                                        "type": "string",
                                                        "example": "2024"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Wrong request!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "properties": {
                                                "action": {
                                                    "description": "name of the action where error is occurred",
                                                    "type": "string",
                                                    "example": "resourse request"
                                                },
                                                "message": {
                                                    "description": "Error message if incoming parameters is wrong",
                                                    "type": "string",
                                                    "example": "Unauthorized"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Resource not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "properties": {
                                                "action": {
                                                    "description": "name of the action where error is occurred",
                                                    "type": "string",
                                                    "example": "resourse request"
                                                },
                                                "message": {
                                                    "description": "Error message if incoming parameters is wrong",
                                                    "type": "string",
                                                    "example": "resource not found"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Content",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "properties": {
                                                "action": {
                                                    "description": "name of the action where error is occurred",
                                                    "type": "string",
                                                    "example": "Get calendar data"
                                                },
                                                "message": {
                                                    "description": "Error message if incoming parameters is wrong",
                                                    "type": "string",
                                                    "example": "Fail due to some errors"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}